m09_12_01/pyproject.toml
[tool.poetry]
name = "m09-12-01"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m09_12_01"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
sqlalchemy = "^2.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.21.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
